Use a list over integers for representing range data
The conversion from integers to floats happens in the publisher. Use integers
until that point.

+++ android_hokuyo/src/org/ros/android/hokuyo/Decoder.java
@@ -32,12 +32,12 @@ class Decoder {
     return result;
   }
 
-  public static List<Float> decode(String buffer, int blockSize) {
+  public static List<Integer> decode(String buffer, int blockSize) {
     Preconditions.checkArgument(buffer.length() % blockSize == 0);
-    List<Float> data = Lists.newArrayList();
+    List<Integer> data = Lists.newArrayList();
     for (int i = 0; i < buffer.length(); i += blockSize) {
       // sensor_msgs/LaserScan uses floats for ranges.
-      data.add((float) decodeValue(buffer.substring(i, i + 3), blockSize));
+      data.add((int) decodeValue(buffer.substring(i, i + 3), blockSize));
     }
     return data;
   }
